Support empty lines in code fences
Previously the paragraph detector included the code fences in the
paragraph up to the first empty line. This worked for trivial cases, but
broke an many real world cases.

I introduce regexes to to this project in this PR. Hopefully, this does
not bring any meaningful downsides I don't know about.